Moved AFlutterTaskIdentity to module aflutter
