+++ src/auto_flutter_dig/task/identity/__init__.py
@@ -1,4 +1,4 @@
-from .aflutter import AflutterTaskIdentity
+from ...module.aflutter.identity import AflutterTaskIdentity
 from .firebase import FirebaseTaskIdentity
 from .flutter import FlutterTaskIdentity
